requirements.txt
asgiref==3.6.0
Django==3.2.16
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Pillow==9.3.0
psycopg2-binary==2.8.6
PyJWT==2.6.0
pytz==2023.3
requests==2.26.0
sqlparse==0.4.4
python-dotenv==0.21.1
importlib-metadata==4.12.0
drf-extra-fields==3.4.0